Update to version 2.0.9

+++ bin/kvstore_restore.py
@@ -4,7 +4,7 @@
 # Enables restore of backed up KV Store from json or json.gz files
 
 # Author: J.R. Murray <jr.murray@deductiv.net>
-# Version: 2.0.8
+# Version: 2.0.9
 
 from __future__ import print_function
 from builtins import str
@@ -16,8 +16,9 @@ import json
 import time
 import glob
 import re
+import gzip
 import kv_common as kv
-from deductiv_helpers import setup_logger, eprint
+from deductiv_helpers import setup_logger, get_uncompressed_size, search_console
 from splunk.clilib import cli_common as cli
 import splunk.rest as rest
 
@@ -26,7 +27,7 @@ sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', '
 from splunklib.searchcommands import \
     dispatch, GeneratingCommand, Configuration, Option, validators
 
-@Configuration()
+@Configuration(distributed=False, type='reporting')
 class KVStoreRestoreCommand(GeneratingCommand):
 	""" %(synopsis)
 
@@ -57,34 +58,29 @@ class KVStoreRestoreCommand(GeneratingCommand):
 		try:
 			cfg = cli.getConfStanza('kvstore_tools','settings')
 		except BaseException as e:
-			eprint("Could not read configuration: " + repr(e))
+			self.write_error("Could not read configuration: " + repr(e))
+			exit(1)
 		
 		# Facility info - prepended to log lines
 		facility = os.path.basename(__file__)
 		facility = os.path.splitext(facility)[0]
-		try:
-			logger = setup_logger(cfg["log_level"], 'kvstore_tools.log', facility)
-		except BaseException as e:
-			eprint("Could not create logger: " + repr(e))
-			exit(1)
-
+		logger = setup_logger(cfg["log_level"], 'kvstore_tools.log', facility)
+		ui = search_console(logger, self)
 		logger.info('Script started by %s' % self._metadata.searchinfo.username)
 
 		session_key = self._metadata.searchinfo.session_key
 		splunkd_uri = self._metadata.searchinfo.splunkd_uri
 
 		# Check for permissions to run the command
-		content = rest.simpleRequest('/services/authentication/current-context?output_mode=json', sessionKey=session_key, method='GET')[1]
+		content = rest.simpleRequest('/services/authentication/current-context?output_mode=json', sessionKey=session_key)[1]
 		content = json.loads(content)
 		current_user = self._metadata.searchinfo.username
 		current_user_capabilities = content['entry'][0]['content']['capabilities']
 		if 'run_kvstore_restore' in current_user_capabilities or 'run_kvst_all' in current_user_capabilities or current_user == 'splunk-system-user':
 			logger.debug("User %s is authorized." % current_user)
 		else:
-			logger.error("User %s is unauthorized. Has the run_kvstore_restore capability been granted?" % current_user)
-			yield({'Error': 'User %s is unauthorized. Has the run_kvstore_restore capability been granted?' % current_user })
-			sys.exit(3)
-
+			ui.exit_error("User %s is unauthorized. Has the run_kvstore_restore capability been granted?" % current_user)
+		
 		# Sanitize input
 		if self.filename:
 			logger.debug('Restore filename: %s' % self.filename)
@@ -114,13 +110,13 @@ class KVStoreRestoreCommand(GeneratingCommand):
 				backup_file_list.append(name)
 
 			if len(backup_file_list) == 0:
+				# Check again in the default path
 				self.filename = os.path.join(default_path, self.filename)
 				for name in glob.glob(self.filename):
 					backup_file_list.append(name)
 
 			if len(backup_file_list) == 0:
-				logger.critical("No matching files: %s" % self.filename)
-				sys.exit(1)
+				ui.exit_error("No matching files: %s" % self.filename)
 		else:
 			logger.debug('No wildcard string found in %s' % self.filename)
 			if os.path.isfile(self.filename):
@@ -128,8 +124,7 @@ class KVStoreRestoreCommand(GeneratingCommand):
 			elif os.path.isfile(os.path.join(default_path, self.filename)):
 				backup_file_list.append(os.path.join(default_path, self.filename))
 			else:
-				logger.critical("File does not exist: %s" % self.filename)
-				sys.exit(1)
+				ui.exit_error("File does not exist: %s" % self.filename)
 
 		deleted_collections = []
 
@@ -143,9 +138,7 @@ class KVStoreRestoreCommand(GeneratingCommand):
 				file_param = matches.group(2)
 				name_split = file_param.split('#')
 			except BaseException as e:
-				logger.critical('Invalid filename: %s\n\t%s' % (name, repr(e)))
-				yield({'Error': 'Invalid filename: %s' % name})
-				sys.exit(1)
+				ui.exit_error('Invalid filename: %s\n\t%s' % (name, repr(e)))
 
 			# Open the file if it's a supported format
 			if (name.endswith('.json') or name.endswith('.json.gz')) and len(name_split)==3:
@@ -153,30 +146,36 @@ class KVStoreRestoreCommand(GeneratingCommand):
 				# Extract the app name and collection name from the file name
 				file_app = name_split[0]
 				file_collection = name_split[1]
-
+				if name.endswith('.json.gz'):
+					data_bytes = get_uncompressed_size(name)
+				else:		
+					data_bytes = os.stat(name).st_size
+				
 				if list_only:
-					yield {'filename': name, 'app': file_app, 'collection': file_collection, 'status': 'ready' }
+					status = 'ready' if data_bytes > 0 else 'empty'
+					yield {'filename': name, 'app': file_app, 'collection': file_collection, 'bytes': data_bytes, 'status': status }
 				else:
-					if not self.append:
-						# Delete the collection contents using the KV Store REST API
+					if data_bytes > 0:
+						if not self.append:
+							# Delete the collection contents using the KV Store REST API
+							try:
+								collection_id = file_app + "/" + file_collection
+								# Make sure we aren't trying to delete the same collection twice
+								if not collection_id in deleted_collections:
+									kv.delete_collection(logger, splunkd_uri, session_key, file_app, file_collection)
+									deleted_collections.append(collection_id)
+							except BaseException as e:
+								ui.exit_error('Failed to delete collection %s/%s: %s' % (file_app, file_collection, repr(e)))
+						
+						# Upload the collection to the KV Store REST API
 						try:
-							collection_id = file_app + "/" + file_collection
-							# Make sure we aren't trying to delete the same collection twice
-							if not collection_id in deleted_collections:
-								kv.delete_collection(logger, splunkd_uri, session_key, file_app, file_collection)
-								deleted_collections.append(collection_id)
+							result, message, record_count = kv.upload_collection(logger, splunkd_uri, session_key, file_app, file_collection, name)
+							yield({ 'filename': name, 'app': file_app, 'collection': file_collection, 'result': result, 'message': message, 'records': record_count })
 						except BaseException as e:
-							logger.critical(repr(e), exc_info=True)
-							yield({'Error': 'Failed to delete collection %s/%s: %s' % (file_app, file_collection, repr(e))})
-							sys.exit(4)
-					
-					# Upload the collection to the KV Store REST API
-					try:
-						result, message, record_count = kv.upload_collection(logger, splunkd_uri, session_key, file_app, file_collection, name)
-						yield({ 'result': result, 'message': message, 'records': record_count })
-					except BaseException as e:
-						logger.error("Error restoring collection: %s" % repr(e), exc_info=True)
-						yield({'result': 'error', 'message': 'Failed to delete collection: %s' % repr(e), 'records': 0})
+							logger.error("Error restoring collection: %s" % repr(e), exc_info=True)
+							yield({ 'filename': name, 'app': file_app, 'collection': file_collection, 'result': 'error', 'message': 'Failed to delete collection: %s' % repr(e), 'records': 0})
+					else:
+						yield({ 'filename': name, 'app': file_app, 'collection': file_collection, 'result': 'skipped', 'message': f'Restored 0 records to {file_app}/{file_collection}', 'records': 0 })
 
 			elif name.endswith('.tar.gz') or name.endswith('.tgz'):
 				logger.info('Skipping filename (unsupported format): %s' % name)
